Redid the GUIs
Kept the GUI forms and added the new screens. Currently the number of players doesn't affect anything nor do the clear form buttons work. The working buttons are ones that change the screens from the history to the calculator. The table that would show probability doesn't display anything, but I did have it set up to take in a string and display that string.