Add files via upload
Added a Deck class to simulate a real deck of cards being used and a GameMain class with a standalone generic gui I used to develop the game until the main GUI is functional. 

Also need to upload the folder with all the card pictures, but I don't want all the images to be in the main calculator project folder.