GUI Changes/Data saving
Redid the Calculator Panel. Set all of the add player buttons to invisible making the number of players Combobox the only way to add players to the table. Put in a diary entry method that will place the data from the table into strings. Code need to add it to the database. Code still needed to populate the history table with information from the database.