Functionaity ready
I have created the Poker Evaluator class that evaluates a the users hand and determines its value whether it is a royal flush or pair, etc. It will be used by the PokerChart class which has most/common poker odds that exist in the game. The poker calculator gets information from the user about its hand and community board and tries to then calculates the probability using outs, loss, and  unkowns and turning it into a ratio and then converts that into a percentage which is our output. Tostring method will have output. but must call calculations method to get it.