Add files via upload
Renamed ComboBox options to match the Suit and Rank enums to prevent the need to convert to and from a string multiple times. Added an ArrayList<Cards> to read and store the combobox selected cards for two players and the five community cards (three flop cards, turn, and river). 
Applied some random values to the JTable used to display odds to show that it will work. 
Created a Logger, but have not finalized where it is needed.