Variables setup Layout Manager Changed
PokerPanel now correctly access the hand, game, player, rank, and suit classes. makeDiaryEntry() method needs to make sure that it correctly is accessing the database. Hand constructor currently in the playerPanelFactory has an issue with the array of cards that will require a relook. Removed the gridbagLayout manager for the bulk of the components and changed it to Spring. The container holding the that panel may need some changes done to correctly readjust on it size.